Refactor the code in peparation for extracting binaries as crates
