+++ Cargo.toml
@@ -13,9 +13,11 @@ waymark-core-backend = { path = "crates/lib/core-backend" }
 waymark-dag = { path = "crates/lib/dag" }
 waymark-garbage-collector-backend = { path = "crates/lib/garbage-collector-backend" }
 waymark-integration-support = { path = "crates/lib/integration-support" }
+waymark-ir-conversions = { path = "crates/lib/ir-conversions" }
 waymark-ir-parser = { path = "crates/lib/ir-parser" }
 waymark-observability = { path = "crates/lib/observability" }
 waymark-observability-macros = { path = "crates/lib/observability-macros" }
+waymark-observability-setup = { path = "crates/lib/observability-setup" }
 waymark-proto = { path = "crates/lib/proto" }
 waymark-runner = { path = "crates/lib/runner" }
 waymark-runner-state = { path = "crates/lib/runner-state" }
@@ -26,11 +28,13 @@ waymark-webapp-backend = { path = "crates/lib/webapp-backend" }
 waymark-webapp-core = { path = "crates/lib/webapp-core" }
 waymark-worker-status-backend = { path = "crates/lib/worker-status-backend" }
 waymark-workflow-registry-backend = { path = "crates/lib/workflow-registry-backend" }
+waymark-smoke-sources = { path = "crates/lib/smoke-sources" }
 
 anyhow = "1"
 async-trait = "0.1"
 chrono = { version = "0.4", default-features = false }
 clap = "4.5"
+console-subscriber = "0.5"
 cron = "0.12"
 proptest = "1.9"
 prost = "0.12"
@@ -49,4 +53,6 @@ tokio = "1"
 tonic = "0.11"
 tonic-build = "0.11"
 tracing = "0.1"
+tracing-chrome = "0.7"
+tracing-subscriber = "0.3"
 uuid = "1"
