+++ crates/lib/ir-conversions/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "waymark-ir-conversions"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+serde_json = { workspace = true }
+waymark-proto = { workspace = true }
