+++ crates/lib/observability-setup/src/common.rs
@@ -1,0 +1,5 @@
+#[derive(Clone, Debug, Default)]
+pub struct ObservabilityOptions {
+    pub console: bool,
+    pub trace_path: Option<String>,
+}
