+++ crates/waymark/Cargo.toml
@@ -19,8 +19,10 @@ path = "src/bin/smoke.rs"
 [dependencies]
 waymark-core-backend = { workspace = true }
 waymark-dag = { workspace = true }
+waymark-ir-conversions = { workspace = true }
 waymark-ir-parser = { workspace = true }
 waymark-observability = { workspace = true }
+waymark-observability-setup = { workspace = true }
 waymark-proto = { workspace = true, features = ["serde", "client", "server"] }
 waymark-runner = { workspace = true }
 waymark-runner-state = { workspace = true }
@@ -36,6 +38,7 @@ waymark-backend-postgres-migrations = { workspace = true }
 waymark-workflow-registry-backend = { workspace = true }
 waymark-worker-status-backend = { workspace = true }
 waymark-backend-memory = { workspace = true }
+waymark-smoke-sources = { workspace = true }
 
 anyhow = "1"
 axum = "0.8"
@@ -70,7 +73,7 @@ regex = "1"
 console-subscriber = { version = "0.5", optional = true }
 
 [features]
-trace = []
+trace = ["waymark-observability-setup/trace"]
 observability = ["trace", "dep:console-subscriber"]
 
 [dev-dependencies]
