+++ crates/waymark/src/bin/waymark-bridge.rs
@@ -39,6 +39,7 @@ use waymark_core_backend::{
     QueuedInstance, QueuedInstanceBatch,
 };
 use waymark_dag::convert_to_dag;
+use waymark_ir_conversions::literal_from_json_value;
 use waymark_runner_state::RunnerState;
 use waymark_scheduler_backend::SchedulerBackend as _;
 use waymark_scheduler_core::{CreateScheduleParams, ScheduleId, ScheduleStatus, ScheduleType};
@@ -1170,7 +1171,7 @@ fn build_queued_instance(
     if let Some(context) = initial_context {
         let inputs = workflow_arguments_to_json_map(&context);
         for (name, value) in inputs {
-            let expr = literal_from_value(&value);
+            let expr = literal_from_json_value(&value);
             let label = format!("input {name} = {value}");
             let _ = state
                 .record_assignment(vec![name.clone()], &expr, None, Some(label))
@@ -1210,67 +1211,6 @@ fn workflow_arguments_to_json_map(args: &proto::WorkflowArguments) -> HashMap<St
         }
     }
     map
-}
-
-fn literal_from_value(value: &Value) -> ir::Expr {
-    match value {
-        Value::Bool(value) => ir::Expr {
-            kind: Some(ir::expr::Kind::Literal(ir::Literal {
-                value: Some(ir::literal::Value::BoolValue(*value)),
-            })),
-            span: None,
-        },
-        Value::Number(number) => {
-            if let Some(value) = number.as_i64() {
-                ir::Expr {
-                    kind: Some(ir::expr::Kind::Literal(ir::Literal {
-                        value: Some(ir::literal::Value::IntValue(value)),
-                    })),
-                    span: None,
-                }
-            } else {
-                ir::Expr {
-                    kind: Some(ir::expr::Kind::Literal(ir::Literal {
-                        value: Some(ir::literal::Value::FloatValue(
-                            number.as_f64().unwrap_or(0.0),
-                        )),
-                    })),
-                    span: None,
-                }
-            }
-        }
-        Value::String(value) => ir::Expr {
-            kind: Some(ir::expr::Kind::Literal(ir::Literal {
-                value: Some(ir::literal::Value::StringValue(value.clone())),
-            })),
-            span: None,
-        },
-        Value::Array(items) => ir::Expr {
-            kind: Some(ir::expr::Kind::List(ir::ListExpr {
-                elements: items.iter().map(literal_from_value).collect(),
-            })),
-            span: None,
-        },
-        Value::Object(map) => {
-            let entries = map
-                .iter()
-                .map(|(key, value)| ir::DictEntry {
-                    key: Some(literal_from_value(&Value::String(key.clone()))),
-                    value: Some(literal_from_value(value)),
-                })
-                .collect();
-            ir::Expr {
-                kind: Some(ir::expr::Kind::Dict(ir::DictExpr { entries })),
-                span: None,
-            }
-        }
-        Value::Null => ir::Expr {
-            kind: Some(ir::expr::Kind::Literal(ir::Literal {
-                value: Some(ir::literal::Value::IsNone(true)),
-            })),
-            span: None,
-        },
-    }
 }
 
 fn proto_schedule_type(value: i32) -> Option<ScheduleType> {
