+++ crates/waymark/src/lib.rs
@@ -3,7 +3,6 @@
 pub mod config;
 pub mod garbage_collector;
 pub mod messages;
-pub mod observability;
 pub mod pool_status;
 pub mod scheduler;
 pub mod server_worker;
@@ -14,7 +13,6 @@ pub mod workers;
 // Worker infrastructure (preserved from the legacy Rust core).
 pub use garbage_collector::{GarbageCollectorConfig, GarbageCollectorTask};
 pub use messages::{MessageError, ast as ir_ast, proto, workflow_argument_value_to_json};
-pub use observability::obs;
 pub use pool_status::{PoolTimeSeries, TimeSeriesEntry, TimeSeriesJsonEntry};
 pub use scheduler::{SchedulerConfig, SchedulerTask};
 pub use server_worker::{WorkerBridgeChannels, WorkerBridgeServer};
