+++ crates/waymark/src/scheduler/task.rs
@@ -10,10 +10,10 @@ use serde_json::Value;
 use tracing::{debug, error, info};
 use uuid::Uuid;
 use waymark_core_backend::QueuedInstance;
+use waymark_ir_conversions::literal_from_json_value;
 use waymark_scheduler_core::{ScheduleId, WorkflowSchedule};
 
 use crate::messages;
-use crate::messages::ast as ir;
 use waymark_dag::DAG;
 
 #[derive(Clone)]
@@ -203,67 +203,6 @@ where
         );
 
         Ok(())
-    }
-}
-
-fn literal_from_json_value(value: &Value) -> ir::Expr {
-    match value {
-        Value::Bool(value) => ir::Expr {
-            kind: Some(ir::expr::Kind::Literal(ir::Literal {
-                value: Some(ir::literal::Value::BoolValue(*value)),
-            })),
-            span: None,
-        },
-        Value::Number(number) => {
-            if let Some(value) = number.as_i64() {
-                ir::Expr {
-                    kind: Some(ir::expr::Kind::Literal(ir::Literal {
-                        value: Some(ir::literal::Value::IntValue(value)),
-                    })),
-                    span: None,
-                }
-            } else {
-                ir::Expr {
-                    kind: Some(ir::expr::Kind::Literal(ir::Literal {
-                        value: Some(ir::literal::Value::FloatValue(
-                            number.as_f64().unwrap_or(0.0),
-                        )),
-                    })),
-                    span: None,
-                }
-            }
-        }
-        Value::String(value) => ir::Expr {
-            kind: Some(ir::expr::Kind::Literal(ir::Literal {
-                value: Some(ir::literal::Value::StringValue(value.clone())),
-            })),
-            span: None,
-        },
-        Value::Array(items) => ir::Expr {
-            kind: Some(ir::expr::Kind::List(ir::ListExpr {
-                elements: items.iter().map(literal_from_json_value).collect(),
-            })),
-            span: None,
-        },
-        Value::Object(map) => {
-            let entries = map
-                .iter()
-                .map(|(key, value)| ir::DictEntry {
-                    key: Some(literal_from_json_value(&Value::String(key.clone()))),
-                    value: Some(literal_from_json_value(value)),
-                })
-                .collect();
-            ir::Expr {
-                kind: Some(ir::expr::Kind::Dict(ir::DictExpr { entries })),
-                span: None,
-            }
-        }
-        Value::Null => ir::Expr {
-            kind: Some(ir::expr::Kind::Literal(ir::Literal {
-                value: Some(ir::literal::Value::IsNone(true)),
-            })),
-            span: None,
-        },
     }
 }
