--- crates/waymark/src/waymark_core/cli/mod.rs
@@ -1,4 +1,0 @@
-//! CLI entrypoints for the Waymark core runtime.
-
-pub mod benchmark;
-pub mod smoke;
