+++ crates/waymark/src/waymark_core/runloop.rs
@@ -23,11 +23,11 @@ use waymark_core_backend::{
 use waymark_workflow_registry_backend::WorkflowRegistryBackend;
 
 use crate::messages::ast as ir;
-use crate::observability::obs;
 use crate::waymark_core::commit_barrier::{CommitBarrier, DeferredInstanceEvent};
 use crate::waymark_core::lock::{InstanceLockTracker, spawn_lock_heartbeat};
 use crate::workers::{ActionCompletion, ActionRequest, BaseWorkerPool, WorkerPoolError};
 use waymark_dag::{DAG, DAGNode, OutputNode, ReturnNode, convert_to_dag};
+use waymark_observability::obs;
 use waymark_runner::synthetic_exceptions::{
     SyntheticExceptionType, build_synthetic_exception_value,
 };
